Makefiles: adjust both to create executables with ".out" suffix for easier cleanup.
These just got hacked together in a minute or two, so they need to be revisited.  The top level directory should have control (which it doesn't) and these should use some smart Make vars to automate entry of the individual C files (and their corresponding object code).  Should be maintenance free for me in the future.